move mac platform to system state

+++ engine/src/orb/core/logger.h
@@ -24,7 +24,7 @@ typedef enum orb_log_level {
     LOG_LEVEL_MAX_ENUM,
 } orb_log_level;
 
-b8 orb_logger_init(u64 *memory_requirement, void *memory);
+b8 orb_logger_init(usize *memory_requirement, void *memory);
 void orb_logger_shutdown();
 
 ORB_API void orb_log(orb_log_level level, const char *message, ...);
